done converting social api call to api-route
