+++ package.json
@@ -31,7 +31,6 @@
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
     "@ensdomains/ens-avatar": "^0.1.6",
-    "@everipedia/iq-utils": "^2.4.0",
     "@everipedia/wagmi-magic-connector": "^1.0.0",
     "@metamask/detect-provider": "^1.2.0",
     "@radix-ui/react-dropdown-menu": "^2.1.2",
